rollback to quast 2.0

+++ quast_tool/libs_/genemark.py
@@ -1,10 +1,11 @@
 ############################################################################
-# Copyright (c) 2011-2012 Saint-Petersburg Academic University
+# Copyright (c) 2011-2013 Saint-Petersburg Academic University
 # All Rights Reserved
 # See file LICENSE for details.
 ############################################################################
-from __future__ import with_statement
 
+from __future__ import with_statement
+import logging
 import os
 import platform
 import shutil
@@ -12,8 +13,9 @@ import subprocess
 import tempfile
 
 from libs import reporting
+from libs import qconfig
 from libs.fastaparser import read_fasta, write_fasta
-from qutils import id_to_str
+from qutils import id_to_str, print_timestamp
 
 def gc_content(sequence):
     GC_count = sequence.count('G') + sequence.count('C')
@@ -68,24 +70,19 @@ def parse_gmhmm_out(out_path):
                     seq.append(line.strip())
 
 
-def add_genes_to_gff(genes, gff_path, gff_header):
+def add_genes_to_gff(genes, gff_path):
     gff = open(gff_path, 'w')
-    gff.write(gff_header)
+    gff.write('##gff out for GeneMark.hmm PROKARYOTIC\n')
+    gff.write('##gff-version 3\n')
 
     for id, gene in enumerate(genes):
         contig_id, strand, left_index, right_index, str_seq = gene
-        length = right_index - left_index
-        gene_id = '>gene_%d|GeneMark.hmm|%d_nt|%s|%d|%d|%s' % (
-            id + 1, length, strand, left_index, right_index, contig_id
-        )
-
-        gff.write(
-            '%s       .      gene    %d %d .     %s       .       ID=%d\n' %
-            (gene_id, left_index, right_index, strand, id + 1))
+        gff.write('%s\tGeneMark\tgene\t%d\t%d\t.\t%s\t.\tID=%d\n' %
+            (contig_id, left_index, right_index, strand, id + 1))
     gff.close()
 
 
-def add_genes_to_fasta(genes, fasta_path, fasta_header):
+def add_genes_to_fasta(genes, fasta_path):
     def inner():
         for id, gene in enumerate(genes):
             contig_id, strand, left_index, right_index, gene_fasta = gene
@@ -95,23 +92,23 @@ def add_genes_to_fasta(genes, fasta_path, fasta_header):
             )
             yield gene_id, gene_fasta
 
-    write_fasta(fasta_path, inner(), header=fasta_header)
+    write_fasta(fasta_path, inner())
 
 
-def gmhmm_p_everyGC(tool_dir, fasta_path, out_name, gene_lengths, err_path):
+def gmhmm_p_everyGC(tool_dir, fasta_path, out_name, gene_lengths, err_path, tmp_dir):
     tool_exec = os.path.join(tool_dir, 'gmhmmp')
     heu_dir = os.path.join(tool_dir, 'heuristic_mod')
 
-    out_gff_path = out_name + '.gff'
-    out_fasta_path = out_name + '_genes.fasta'
+    out_gff_path = out_name + '_genes.gff'
+    #out_fasta_path = out_name + '_genes.fasta'
 
-    work_dir = tempfile.mkdtemp()
+    work_dir = tempfile.mkdtemp(dir=tmp_dir)
     for id, seq in read_fasta(fasta_path):
         gc = min(70, max(30, gc_content(seq)))
         curr_filename = str(gc - gc % 5) + '.fasta'
         current_path = os.path.join(work_dir, curr_filename)
         with open(current_path, 'a') as curr_out:
-            curr_out.write(id + '\n' + seq)
+            curr_out.write('>' + id + '\n' + seq + '\n')
 
     genes = []
     _ , _, file_names = os.walk(work_dir).next()
@@ -124,66 +121,72 @@ def gmhmm_p_everyGC(tool_dir, fasta_path, out_name, gene_lengths, err_path):
             if gmhmm_p(tool_exec, file_path, heu_path, file_out, err_file):
                 genes.extend(parse_gmhmm_out(file_out))
 
-    shutil.rmtree(work_dir)
-    gff_header = '''
-##gff out for GeneMark.hmm PROKARYOTIC
-##Sequence file name: %s
-''' % fasta_path
-    add_genes_to_gff(genes, out_gff_path, gff_header)
-    fasta_header = '''
-##fasta out for GeneMark.hmm PROKARYOTIC
-##Sequence file name: %s
-''' % fasta_path
-    add_genes_to_fasta(genes, out_fasta_path, fasta_header)
+    if not qconfig.debug:
+        shutil.rmtree(work_dir)
+
+    add_genes_to_gff(genes, out_gff_path)
+    #add_genes_to_fasta(genes, out_fasta_path)
 
     cnt = [sum([gene[3] - gene[2] > x for gene in genes]) for x in gene_lengths]
     unique_count = len(set([gene[4] for gene in genes]))
     total_count = len(genes)
 
-    return out_gff_path, out_fasta_path, unique_count, total_count, cnt
+    #return out_gff_path, out_fasta_path, unique_count, total_count, cnt
+    return out_gff_path, unique_count, total_count, cnt
 
 
-def do(fasta_pathes, gene_lengths, out_dir, lib_dir):
+def predict_genes(id, fasta_path, gene_lengths, out_dir, tool_dir, tmp_dir):
+    log = logging.getLogger('quast')
+    log.info('  ' + id_to_str(id) + os.path.basename(fasta_path))
 
-    print 'Running GeneMark tool...'
+    out_name = os.path.basename(fasta_path)
+    out_path = os.path.join(out_dir, out_name)
+    err_path = os.path.join(out_dir, out_name + '_genemark.stderr')
+    #out_gff_path, out_fasta_path, unique, total, cnt = gmhmm_p_everyGC(tool_dir,
+    #    fasta_path, out_path, gene_lengths, err_path)
+    out_gff_path, unique, total, cnt = gmhmm_p_everyGC(tool_dir,
+        fasta_path, out_path, gene_lengths, err_path, tmp_dir)
 
+    log.info('  ' + id_to_str(id) + '  Genes = ' + str(unique) + ' unique, ' + str(total) + ' total')
+    log.info('  ' + id_to_str(id) + '  Predicted genes (GFF): ' + out_gff_path)
+
+    log.info('  ' + id_to_str(id) + 'Gene prediction is finished.')
+    return unique, cnt
+
+
+def do(fasta_paths, gene_lengths, out_dir):
+    print_timestamp()
+    log = logging.getLogger('quast')
+    log.info('Running GeneMark tool...')
+
+    tmp_dir = os.path.join(out_dir, 'tmp')
     if not os.path.isdir(out_dir):
         os.mkdir(out_dir)
+    if not os.path.isdir(tmp_dir):
+        os.mkdir(tmp_dir)
 
     if platform.system() == 'Darwin':
-        tool_dir = os.path.join(lib_dir, 'genemark_suite_macosx/gmsuite')
+        tool_dir = os.path.join(qconfig.LIBS_LOCATION, 'genemark', 'macosx')
     else:
         if platform.architecture()[0] == '64bit':
-            tool_dir  = os.path.join(lib_dir, 'genemark_suite_linux_64/gmsuite')
+            tool_dir  = os.path.join(qconfig.LIBS_LOCATION, 'genemark', 'linux_64')
         else:
-            tool_dir  = os.path.join(lib_dir, 'genemark_suite_linux_32/gmsuite')
+            tool_dir  = os.path.join(qconfig.LIBS_LOCATION, 'genemark', 'linux_32')
 
     install_genemark(tool_dir)
 
-    for id, fasta_path in enumerate(fasta_pathes):
+    n_jobs = min(len(fasta_paths), qconfig.max_threads)
+    from joblib import Parallel, delayed
+    results = Parallel(n_jobs=n_jobs)(delayed(predict_genes)(id, fasta_path, gene_lengths, out_dir, tool_dir, tmp_dir)
+        for id, fasta_path in enumerate(fasta_paths))
+
+    # saving results
+    for id, fasta_path in enumerate(fasta_paths):
         report = reporting.get(fasta_path)
+        unique, cnt = results[id]
+        report.add_field(reporting.Fields.PREDICTED_GENES_UNIQUE, unique)
+        report.add_field(reporting.Fields.PREDICTED_GENES, cnt)
 
-        print ' ', id_to_str(id), os.path.basename(fasta_path),
-
-        out_name = os.path.basename(fasta_path)
-        out_path = os.path.join(out_dir, out_name)
-        err_path = os.path.join(out_dir, 'genemark.errors')
-        out_gff_path, out_fasta_path, unique, total, cnt = gmhmm_p_everyGC(tool_dir,
-            fasta_path, out_path, gene_lengths, err_path)
-
-        print ', Genes =', unique, 'unique,', total, 'total'
-        print '    GeneMark output: %s and %s'%(out_gff_path, out_fasta_path)
-
-        report.add_field(reporting.Fields.GENEMARKUNIQUE, unique)
-        report.add_field(reporting.Fields.GENEMARK, cnt)
-
-        print '  Done'
-
-
-if __name__ == '__main__':
-    fasta_pathes = ['../test_data/assembly_10K_1.fasta', '../test_data/assembly_10K_2.fasta']
-    print fasta_pathes
-    out_dir = '../run_test_data/genemark_out'
-    lib_dir = ''
-    gene_lengths = [100, 1000, 10000]
-    do(fasta_pathes, gene_lengths, out_dir, lib_dir)
+    if not qconfig.debug:
+        shutil.rmtree(tmp_dir)
+    log.info('  Done.')
